Unify logic for running parser tests from external suites (#10444)
