+++ package.json
@@ -60,6 +60,7 @@
     "lerna-changelog": "^0.5.0",
     "lint-staged": "^9.2.0",
     "lodash": "^4.17.13",
+    "mergeiterator": "^1.2.5",
     "output-file-sync": "^2.0.0",
     "prettier": "^1.19.1",
     "pump": "^3.0.0",
