--- scripts/tests/flow/flow_tests_whitelist.txt
@@ -1,32 +1,0 @@
-# This file lists tests that are known to produce incorrect results when parsed
-# with the babel parser:
-#
-# - Tests that are expected to parse successfully but for which the babel parser reports
-#   a syntax error
-# - Tests that contain invalid syntax but for which the babel parser reports no syntax
-#   error
-#
-# Entries should be removed incrementally as the babel parser is improved.
-
-JSX_invalid/migrated_0000.js
-arrow_function_invalid/migrated_0002.js
-async_await/migrated_0007.js
-async_await/migrated_0020.js
-async_await/migrated_0024.js
-async_await/migrated_0027.js
-async_generators/migrated_0007.js
-class_properties/migrated_0021.js
-class_properties/migrated_0026.js
-decorators/migrated_0003.js
-private_class_properties/multiple.js
-private_class_properties/getter_and_field.js
-private_class_properties/getter_duplicate.js
-private_class_properties/setter_and_field.js
-private_class_properties/setter_duplicate.js
-types/member/reserved_words.js
-class_method_kinds/polymorphic_getter.js
-ES6/modules/migrated_0020.js
-export_import_reserved_words/migrated_0003.js
-export_statements/export_trailing_comma.js
-nullish_coalescing/precedence_and.js
-nullish_coalescing/precedence_or.js
